# src/main/java/com/ps/weather_prediction/model/response/WeatherResponse.java
package com.ps.weather_prediction.model.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class WeatherResponse {
    String city;
    List<WeatherDetails> weatherDetails;
}